requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
email-validator>=1.1.3

# Data Science and Machine Learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization and Dashboard
streamlit>=1.22.0
plotly>=5.13.0

# Database and Caching
aiosqlite>=0.17.0
redis>=4.0.0
sqlalchemy>=1.4.0
alembic>=1.7.0

# Environment and Configuration
python-dotenv>=0.19.0
pyyaml>=5.4.0

# AI and Language Models
langchain>=0.0.330
langchain-community>=0.0.10
langchain-core>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0

# Monitoring and Metrics
prometheus-client>=0.12.0
python-json-logger>=2.0.0

# AWS Integration
boto3>=1.26.0
botocore>=1.29.0

# Security
cryptography>=3.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
httpx>=0.18.0

# Type Hints
typing-extensions>=4.0.0
